+++ llvm/lib/Target/AMDGPU/SIMachineFunctionInfo.h
@@ -308,6 +308,7 @@ struct SIMachineFunctionInfo final : public yaml::MachineFunctionInfo {
   unsigned ScratchReservedForDynamicVGPRs = 0;
 
   unsigned NumKernargPreloadSGPRs = 0;
+  unsigned KernargPreloadOffset = 0;
 
   unsigned MinNumAGPRs = ~0u;
 
@@ -367,6 +368,7 @@ template <> struct MappingTraits<SIMachineFunctionInfo> {
     YamlIO.mapOptional("scratchReservedForDynamicVGPRs",
                        MFI.ScratchReservedForDynamicVGPRs, 0);
     YamlIO.mapOptional("numKernargPreloadSGPRs", MFI.NumKernargPreloadSGPRs, 0);
+    YamlIO.mapOptional("kernargPreloadOffset", MFI.KernargPreloadOffset, 0);
     YamlIO.mapOptional("isWholeWaveFunction", MFI.IsWholeWaveFunction, false);
     YamlIO.mapOptional("minNumAGPRs", MFI.MinNumAGPRs, ~0u);
   }
@@ -499,6 +501,9 @@ private:
   // Tracks information about user SGPRs that will be setup by hardware which
   // will apply to all wavefronts of the grid.
   GCNUserSGPRUsageInfo UserSGPRInfo;
+
+  // Dword offset in the kernel argument segment where preloading starts.
+  unsigned KernargPreloadOffset = 0;
 
   // Feature bits required for inputs passed in system SGPRs.
   bool WorkGroupIDX : 1; // Always initialized.
@@ -1020,6 +1025,11 @@ public:
 
   unsigned getNumKernargPreloadedSGPRs() const {
     return UserSGPRInfo.getNumKernargPreloadSGPRs();
+  }
+
+  unsigned getKernargPreloadOffset() const { return KernargPreloadOffset; }
+  void setKernargPreloadOffset(unsigned Offset) {
+    KernargPreloadOffset = Offset;
   }
 
   unsigned getNumWaveDispatchSGPRs() const { return NumWaveDispatchSGPRs; }